backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.26.0
langdetect==1.0.9
requests
beautifulsoup4
duckduckgo-search==5.3.1b1
